DASH Live (dynamic MPDs) support Playback of live DASH content with dynamic (reloadable) MPDs
Implementation of DASH Live with dynamic (reloadable MPDs)
This implementation has been tested with following scenarios

Template segment list - Yes
Template timeline (build from duration) - No
List Representation - No
Base Representation - No.

All aforementioned cases may support live playback with reloadable MPDs.
Due to lack of such cases, their implementation is... preliminary.

Added 3 live streams to video clips, including Canal+ Live stream.

Observations:

1.
There are situations where pipeline "stalls" when playing live material.
In such cases there are no timestamps coming up from underlying player.
Issue related with PTS/DTS wraparounds? Has not been investigated yet.

2.
Canal+ Live content. Observed to "stop" when period is over. Expected
scenario in normal situations, however, it should not be the case for
this particular stream. Not investigated yet.

Possible causes:

We are out of data but new MPD has not arrived yet, as such playback
termination is issues (i.e. null packet recieved from demux)

MPD is recieved but has not arrived at DashClient yet. Simple counter
measure for this has been added, but further testing and veficition is
necessary.

+++ JuvoPlayer.Tests/UnitTests/TSMPDParserHelper.cs
@@ -993,7 +993,7 @@ namespace JuvoPlayer.Tests.UnitTests
                 item.MaximumSAPPeriod = input.maximumSAPPeriod;
 
             if (input.startWithSAPSpecified)
-                item.StartWithSAP = (int)input.startWithSAP;
+                item.StartWithSAP = input.startWithSAP;
 
             if (input.maxPlayoutRateSpecified)
                 item.MaxPlayoutRate = input.maxPlayoutRate;
